# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Pets-project-app-backend",
    "description": "This is the API of the pets application."
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
...
    "/notices/category/{category}": {
      "post": {
        "tags": ["Notices"],
        "summary": "Create a new notice",
        "description": "Create a new notice with the provided information",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "category",
            "description": "The category of the notice",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Notice object that needs to be created",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 48,
                  "description": "The title of the notice"
                },
                "name": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 24,
                  "description": "The name of the item"
                },
                "birthdate": {
                  "type": "string",
                  "format": "date",
                  "description": "The birthdate of the item"
                },
                "breed": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 24,
                  "description": "The breed of the item"
                },
                "location": {
                  "type": "string",
                  "description": "The location of the item"
                },
                "comments": {
                  "type": "string",
                  "minLength": 8,
                  "maxLength": 120,
                  "description": "The comments about the item"
                },
                "price": {
                  "type": "number",
                  "minimum": 0,
                  "description": "The price of the item"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully created the notice",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "description": "Title of the dog post"
                    },
                    "name": {
                      "type": "string",
                      "description": "Name of the dog"
                    },
                    "birthdate": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Birthdate of the dog"
                    },
                    "breed": {
                      "type": "string",
                      "description": "Breed of the dog"
                    },
                    "location": {
                      "type": "string",
                      "description": "Location where the dog is available"
                    },
                    "comments": {
                      "type": "string",
                      "description": "Additional comments on the dog"
                    },
                    "price": {
                      "type": "number",
                      "description": "Price of the dog"
                    },
                    "category": {
                      "type": "string",
                      "description": "Category of the dog post (sell/buy/adoption)"
                    },
                    "favorite": {
                      "type": "boolean",
                      "description": "Whether the dog is marked as favorite"
                    },
                    "avatarURL": {
                      "type": "string",
                      "format": "uri",
                      "description": "URL to the dog's avatar"
                    },
                    "owner": {
                      "type": "string",
                      "description": "ID of the owner of the dog"
                    },
                    "_id": {
                      "type": "string",
                      "description": "ID of the dog resource"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp when the dog resource was created"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp when the dog resource was last updated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Notice not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": ["Notices"],
        "summary": "Get notices by category",
        "description": "Get notices filtered by category",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notices"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "notices/{category}/{title}" : {
            "get": {
        "tags": ["Notices"],
        "summary": "Get notices by category and title",
        "description": "Get notices filtered by category and title",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notices"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notices/{noticeId}": {
      "get": {
        "tags": ["Notices"],
        "summary": "Get a notice by its id",
        "operationId": "getNoticeById",
        "parameters": [
          {
            "name": "noticeId",
            "in": "path",
            "description": "Id of the notice to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "minLength": 2,
                      "maxLength": 48
                    },
                    "name": {
                      "type": "string",
                      "minLength": 2,
                      "maxLength": 24
                    },
                    "birthdate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "breed": {
                      "type": "string",
                      "minLength": 2,
                      "maxLength": 24
                    },
                    "location": {
                      "type": "string"
                    },
                    "comments": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 120
                    },
                    "price": {
                      "type": "number",
                      "minimum": 0
                    },
                    "category": {
                      "type": "string"
                    },
                    "favorite": {
                      "type": "boolean"
                    },
                    "gender": {
                      "type": "string"
                    },
                    "avatarURL": {
                      "type": "string"
                    },
                    "owner": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "mobilePhone": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Notice not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Notices"],
        "summary": "Delete a notice",
        "parameters": [
          {
            "name": "noticeId",
            "in": "path",
            "description": "Id of the notice to be deleted",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Notice deleted",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Success message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "Notice not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notices/": {
      "get": {
        "tags": ["Notices"],
        "operationId": "getUserNotices",
        "summary": "Retrieve the user's notices",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "type": "integer",
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "birthdate": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "breed": {
                        "type": "string"
                      },
                      "location": {
                        "type": "string"
                      },
                      "comments": {
                        "type": "string"
                      },
                      "price": {
                        "type": "number"
                      },
                      "category": {
                        "type": "string"
                      },
                      "favorite": {
                        "type": "boolean"
                      },
                      "avatarURL": {
                        "type": "string"
                      },
                      "owner": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notices/{userId}/favorites/{noticeId}": {
      "put": {
        "tags": ["Notices"],
        "summary": "Add or remove a Notice as a favorite",
        "description": "Add or remove a Notice as a favorite for the specified User",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The User ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "noticeId",
            "in": "path",
            "description": "The Notice ID",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "description": "Title of the dog post"
                    },
                    "name": {
                      "type": "string",
                      "description": "Name of the dog"
                    },
                    "birthdate": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Birthdate of the dog"
                    },
                    "breed": {
                      "type": "string",
                      "description": "Breed of the dog"
                    },
                    "location": {
                      "type": "string",
                      "description": "Location where the dog is available"
                    },
                    "comments": {
                      "type": "string",
                      "description": "Additional comments on the dog"
                    },
                    "price": {
                      "type": "number",
                      "description": "Price of the dog"
                    },
                    "category": {
                      "type": "string",
                      "description": "Category of the dog post (sell/buy/adoption)"
                    },
                    "favorite": {
                      "type": "boolean",
                      "description": "Whether the dog is marked as favorite"
                    },
                    "avatarURL": {
                      "type": "string",
                      "format": "uri",
                      "description": "URL to the dog's avatar"
                    },
                    "owner": {
                      "type": "string",
                      "description": "ID of the owner of the dog"
                    },
                    "_id": {
                      "type": "string",
                      "description": "ID of the dog resource"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp when the dog resource was created"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp when the dog resource was last updated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notices/{userid}/favorites": {
      "get": {
        "tags": ["Notices"],
        "summary": "Get a list of favorite pet notices for a user",
        "description": "Get a list of favorite pet notices for a user",
        "parameters": [
          {
            "name": "userid",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of favorite pet notices for the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 48
                      },
                      "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 24
                      },
                      "birthdate": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "breed": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 24
                      },
                      "location": {
                        "type": "string"
                      },
                      "comments": {
                        "type": "string",
                        "minLength": 8,
                        "maxLength": 120
                      },
                      "price": {
                        "type": "number",
                        "minimum": 0
                      },
                      "category": {
                        "type": "string"
                      },
                      "favorite": {
                        "type": "boolean"
                      },
                      "avatarURL": {
                        "type": "string"
                      },
                      "owner": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Notice not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notices/search/{title}": {
      "get": {
        "tags": ["Notices"],
        "summary": "search for notices by title",
        "description": "Endpoint to search for notices by title",
        "operationId": "findNoticeByTitle",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "birthdate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "breed": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  },
                  "comments": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "category": {
                    "type": "string"
                  },
                  "favorite": {
                    "type": "boolean"
                  },
                  "avatarURL": {
                    "type": "string"
                  },
                  "owner": {
                    "type": "string"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notices/avatars/{noticeId}": {
      "patch": {
        "tags": ["Notices"],
        "summary": "Update notice avatar",
        "description": "Update the avatar of the notice with the specified ID",
        "parameters": [
          {
            "name": "noticeId",
            "in": "path",
            "description": "The ID of the notice to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "avatar",
            "in": "formData",
            "description": "The avatar to be uploaded",
            "required": true,
            "type": "file"
          },
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The avatar has been updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "avatarURL": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Notice not found"
          },
          "500": {
            "description": "An internal server error has occurred",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT authorization token"
    }
  },
...
}